Migrated to TpT paradigm. (#88)
* Migrated to TpT paradigm.

* Circle config changes.

+++ src/integration/kotlin/Kafka2HbaseSpec.kt
@@ -7,6 +7,7 @@ import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.AmazonS3ClientBuilder
 import com.beust.klaxon.Klaxon
 import io.kotlintest.shouldBe
+import io.kotlintest.shouldNotBe
 import io.kotlintest.specs.StringSpec
 import lib.*
 import org.apache.kafka.clients.producer.KafkaProducer
@@ -26,26 +27,29 @@ class Kafka2HBaseSpec: StringSpec(){
             val parser = MessageParser()
             val converter = Converter()
             val topic = uniqueTopicName()
-            val startingCounter = waitFor { hbase.getCount(topic) }
-            val s3Client = getS3Client()
-            val summaries = s3Client.listObjectsV2("kafka2s3", "prefix").objectSummaries
-            summaries.forEach{s3Client.deleteObject("kafka2s3", it.key)}
-            val body = wellformedValidPayload()
-            val timestamp = converter.getTimestampAsLong(getISO8601Timestamp())
-            val key = parser.generateKey(converter.convertToJson(getId().toByteArray()))
-            log.info("Sending well-formed record to kafka topic '${String(topic)}'.")
-            producer.sendRecord(topic, "key1".toByteArray(), body, timestamp)
-            log.info("Sent well-formed record to kafka topic '${String(topic)}'.")
+            val matcher = TextUtils().topicNameTableMatcher(topic)
+            matcher shouldNotBe null
+            if (matcher != null) {
+                val namespace = matcher.groupValues[1]
+                val tableName = matcher.groupValues[2]
+                val qualifiedTableName = "$namespace:$tableName".replace("-", "_")
+                hbase.ensureTable(qualifiedTableName)
+                val s3Client = getS3Client()
+                val summaries = s3Client.listObjectsV2("kafka2s3", "prefix").objectSummaries
+                summaries.forEach{s3Client.deleteObject("kafka2s3", it.key)}
+                val body = wellformedValidPayload()
+                val timestamp = converter.getTimestampAsLong(getISO8601Timestamp())
+                val key = parser.generateKey(converter.convertToJson(getId().toByteArray()))
+                log.info("Sending well-formed record to kafka topic '$topic'.")
+                producer.sendRecord(topic.toByteArray(), "key1".toByteArray(), body, timestamp)
+                log.info("Sent well-formed record to kafka topic '$topic'.")
+                val referenceTimestamp = converter.getTimestampAsLong(getISO8601Timestamp())
+                val storedValue = waitFor { hbase.getCellBeforeTimestamp(qualifiedTableName, key, referenceTimestamp) }
+                String(storedValue ?: "null".toByteArray()) shouldBe String(body)
+                val summaries1 = s3Client.listObjectsV2("kafka2s3", "prefix").objectSummaries
+                summaries1.size shouldBe 0
+            }
 
-            val referenceTimestamp = converter.getTimestampAsLong(getISO8601Timestamp())
-            val storedValue = waitFor { hbase.getCellBeforeTimestamp(topic, key, referenceTimestamp) }
-            String(storedValue ?: "null".toByteArray()) shouldBe String(body)
-
-            val counter = waitFor { hbase.getCount(topic) }
-            counter shouldBe startingCounter + 1
-
-            val summaries1 = s3Client.listObjectsV2("kafka2s3", "prefix").objectSummaries
-            summaries1.size shouldBe 0
         }
 
         "Messages with previously received identifiers are written as new versions to hbase but not to dlq" {
@@ -59,12 +63,17 @@ class Kafka2HBaseSpec: StringSpec(){
             val parser = MessageParser()
             val converter = Converter()
             val topic = uniqueTopicName()
-            val startingCounter = waitFor { hbase.getCount(topic) }
-
-            val body1 = wellformedValidPayload()
-            val kafkaTimestamp1 = converter.getTimestampAsLong(getISO8601Timestamp())
+            val matcher1 = TextUtils().topicNameTableMatcher(topic)
+            matcher1 shouldNotBe null
             val key = parser.generateKey(converter.convertToJson(getId().toByteArray()))
-            hbase.putVersion(topic, key, body1, kafkaTimestamp1)
+            val body1 = wellformedValidPayload()
+            if (matcher1 != null) {
+                val namespace = matcher1.groupValues[1]
+                val tableName = matcher1.groupValues[2]
+                val qualifiedTableName = "$namespace:$tableName".replace("-", "_")
+                val kafkaTimestamp1 = converter.getTimestampAsLong(getISO8601Timestamp())
+                hbase.putVersion(qualifiedTableName, key, body1, kafkaTimestamp1)
+            }
 
             Thread.sleep(1000)
             val referenceTimestamp = converter.getTimestampAsLong(getISO8601Timestamp())
@@ -72,19 +81,23 @@ class Kafka2HBaseSpec: StringSpec(){
 
             val body2 = wellformedValidPayload()
             val kafkaTimestamp2 = converter.getTimestampAsLong(getISO8601Timestamp())
-            producer.sendRecord(topic, "key2".toByteArray(), body2, kafkaTimestamp2)
+            producer.sendRecord(topic.toByteArray(), "key2".toByteArray(), body2, kafkaTimestamp2)
 
            val summaries1 = s3Client.listObjectsV2("kafka2s3", "prefix").objectSummaries
             summaries1.size shouldBe 0
 
-            val storedNewValue = waitFor { hbase.getCellAfterTimestamp(topic, key, referenceTimestamp) }
-            storedNewValue shouldBe body2
+            val matcher = TextUtils().topicNameTableMatcher(topic)
+            matcher shouldNotBe null
+            if (matcher != null) {
+                val namespace = matcher.groupValues[1]
+                val tableName = matcher.groupValues[2]
+                val qualifiedTableName = "$namespace:$tableName".replace("-", "_")
+                val storedNewValue = waitFor { hbase.getCellAfterTimestamp(qualifiedTableName, key, referenceTimestamp) }
+                storedNewValue shouldBe body2
 
-            val storedPreviousValue = waitFor { hbase.getCellBeforeTimestamp(topic, key, referenceTimestamp) }
-            storedPreviousValue shouldBe body1
-
-            val counter = waitFor { hbase.getCount(topic) }
-            counter shouldBe startingCounter + 2
+                val storedPreviousValue = waitFor { hbase.getCellBeforeTimestamp(qualifiedTableName, key, referenceTimestamp) }
+                storedPreviousValue shouldBe body1
+            }
         }
 
         "Malformed json messages are written to dlq topic" {
@@ -95,7 +108,7 @@ class Kafka2HBaseSpec: StringSpec(){
             val body = "junk".toByteArray()
             val timestamp = converter.getTimestampAsLong(getISO8601Timestamp())
             val producer = KafkaProducer<ByteArray, ByteArray>(Config.Kafka.producerProps)
-            producer.sendRecord(topic, "key3".toByteArray(), body, timestamp)
+            producer.sendRecord(topic.toByteArray(), "key3".toByteArray(), body, timestamp)
             val malformedRecord = MalformedRecord("key3", String(body), "Invalid json")
             val expected = Klaxon().toJsonString(malformedRecord)
             Thread.sleep(10_000)
@@ -111,13 +124,13 @@ class Kafka2HBaseSpec: StringSpec(){
             val body = """{ "key": "value" } """.toByteArray()
             val timestamp = converter.getTimestampAsLong(getISO8601Timestamp())
             val producer = KafkaProducer<ByteArray, ByteArray>(Config.Kafka.producerProps)
-            producer.sendRecord(topic, "key4".toByteArray(), body, timestamp)
+            producer.sendRecord(topic.toByteArray(), "key4".toByteArray(), body, timestamp)
             Thread.sleep(10_000)
             val s3Object = s3Client.getObject("kafka2s3", "prefix/test-dlq-topic/${SimpleDateFormat("YYYY-MM-dd").format(Date())}/key4").objectContent
             val actual = s3Object.bufferedReader().use(BufferedReader::readText)
             val malformedRecord = MalformedRecord(
                 "key4", String(body),
-                "Invalid schema for key4:${String(topic)}:0:0: Message failed schema validation: '#: required key [message] not found'."
+                "Invalid schema for key4:$topic:0:0: Message failed schema validation: '#: required key [message] not found'."
             )
             val expected = Klaxon().toJsonString(malformedRecord)
             actual shouldBe expected
